+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Latest:
 ### v1.4.3
 **Key Updates**
  - Add `data-pieces` and `parity-pieces` flags to `siac renter upload`
+ - Integrate SiaMux
+ - Initialize defaults for the host's ephemeral account settings
  - Add SIA_DATA_DIR environment variable for setting the data directory for
    siad/siac
  - Made build process deterministic. Moved related scripts into `release-scripts`
@@ -25,6 +27,9 @@ Latest:
    writing to a file after the contractmanager had shutdown
  - Fixed panic where the watchdog would try to write to the contractor's log
    after the contractor had shutdown
+
+**Other**
+ - Upgrade host metadata to v1.4.3
 
 ### v1.4.2.1
 **Key Updates**